fixed setupscript, because it was giving the wrong .vimrc, and a double linusxscript symbolic link was created one level down after second time

+++ setup_user_environment.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -x
 if [ -L $0 ] ; then
     SCRIPT_DIR=$(dirname $(readlink -f $0)) ;
 else
@@ -11,9 +12,15 @@ echo $CALLING_DIR
 
 
 ln -s $SCRIPT_DIR/.bash_aliases $CALLING_DIR/.bash_aliases
-ln -s $SCRIPT_DIR/.vimrc $CALLING_DIR/.vimrc
+ln -s $SCRIPT_DIR/.vimrc_withdebug $CALLING_DIR/.vimrc
 ln -s $SCRIPT_DIR/.screenrc $CALLING_DIR/.screenrc
 ln -s $SCRIPT_DIR/.bashrc $CALLING_DIR/.bashrc
-ln -s /data/scripts/linuxscripts $CALLING_DIR/linuxscripts
 
+if [ ! -e "$CALLING_DIR/linuxscripts" ]
+then
+    ln -s /data/scripts/linuxscripts $CALLING_DIR/linuxscripts
+fi
+set +x
 source $CALLING_DIR/.bashrc
+
+
